requirements.txt
annotated-types==0.6.0
anyio==4.3.0
api-display-purposes==0.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cattrs==23.2.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
clarifai==10.2.1
clarifai-grpc==10.2.1
configparser==3.8.1
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
decorator==4.0.11
EasyProcess==1.1
emoji==2.10.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filetype==1.0.5
flatbuffers==23.5.26
fonttools==4.50.0
future==1.0.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huepy==1.2.1
idna==2.10
imageio==2.1.2
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
inquirerpy==0.3.4
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instagrapi==2.0.3
Js2Py==0.74
jsonschema==2.6.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.3
mdurl==0.1.2
MeaningCloud-python==2.0.0
mock==5.1.0
MouseInfo==0.1.3
moviepy==0.2.3.2
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.10.0
openpyxl==3.1.2
outcome==1.3.0.post0
ozon3==4.0.2
packaging==24.0
pandas==2.2.1
pfzy==0.3.4
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
plyer==2.1.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.20.3
py==1.11.0
PyAutoGUI==0.9.54
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyGetWindow==0.0.9
Pygments==2.17.2
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-rapidjson==1.16
python-telegram-bot==21.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyVirtualDisplay==3.0
PyYAML==6.0.1
ratelimit==2.2.1
regex==2023.12.25
requests==2.31.0
requests-cache==1.2.0
requests-toolbelt==1.0.0
responses==0.25.0
retry==0.9.2
retry-requests==2.0.0
rich==13.7.1
schedule==1.2.1
schema==0.7.5
selenium==4.18.1
semantic-version==2.10.0
setuptools-rust==1.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sseclient-py==1.8.0
tabulate==0.9.0
tomli==2.0.1
tqdm==4.11.2
trio==0.25.0
trio-websocket==0.11.1
tritonclient==2.43.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
url-normalize==1.4.3
urllib3==2.2.1
wcwidth==0.2.13
webdriverdownloader==1.1.0.4
wsproto==1.2.0
